multiple corrections/expansions
migrated pieces of the license to implement an ILicensePart concept which allows for the injection of custom license components.  Changed the signing and verfication methods to work properly.  Changed the serial Number to fit the pattern XXXX-XXXXXX-XXXXXXXX-XXXXXX-XXXX which is built from a hash of the actual application id and a checksum